Refactor/Chore contentview: Create postsViewModel, rename Results to Post for readability, Move Post struct out of ContentView
