--- postApp/postsAPI.swift
@@ -1,18 +1,0 @@
-//
-//  postsAPI.swift
-//  postApp
-//
-//  Created by Matt Wilkes on 30/08/2024.
-//
-
-import Foundation
-import SwiftUI
-
-struct PostData: Identifiable {
-    let userId: Int
-    let id: Int
-    let title: String
-    let body: String
-}
-
-
